refactor(sidebar): extract company selector with react query (#347)
* chore: start pdf unit abbreviation branch

* feat: add pdf unit display formatter

* feat: add unit column to sales invoice pdf

* feat: add unit column to sales invoice pdf

* feat: add unit column to purchase invoice pdf

* test: cover pdf unit abbreviation and unit column

* feat: add product unit fields and decimal inventory migration

* feat: enforce unit and decimal inventory rules in api

* feat: add unit and decimal controls to product and inventory ui

* feat: show product unit in invoice quantity fields

* feat: support decimal invoice quantities per product

* style(pdf): compact invoice table headers for dense columns

* fix: correct invoice pdf header helper indentation

* fix(pdf): format quantity display by product decimal setting

* refactor(sidebar): extract company selector with react query (#346)

+++ backend/src/models/inventory.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 from src.db.base import Base
 
@@ -9,6 +9,6 @@ class Inventory(Base):
     id = Column(Integer, primary_key=True, index=True)
     company_id = Column(Integer, ForeignKey("company_profiles.id"), nullable=True, index=True)
     product_id = Column(Integer, ForeignKey("products.id"), unique=True, nullable=False)
-    quantity = Column(Integer, default=0, nullable=False)
+    quantity = Column(Numeric(12, 3), default=0, nullable=False)
 
     product = relationship("Product")
